BackEnd/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
channels==4.1.0
Django==5.1.1
django-elasticsearch-dsl==7.4
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elastic-transport==8.15.0
elasticsearch==7.17.12
elasticsearch-dsl==7.4.1
packaging==24.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
